requirements.txt
agate==1.9.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arcgis==2.4.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
cryptography==44.0.3
cycler==0.12.1
daff==1.4.2
dask==2024.12.1
dask-expr==1.1.21
db-dtypes==1.4.3
dbt-adapters==1.15.3
dbt-bigquery==1.9.2
dbt-common==1.25.0
dbt-core==1.9.4
dbt-extractor==0.6.0
dbt-protos==1.0.315
dbt-semantic-interfaces==0.7.4
debugpy==1.8.14
decorator==5.2.1
deepdiff==7.0.1
defusedxml==0.7.1
dlt==1.11.0
docstring_parser==0.16
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.58.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.5.1
gcloud==0.18.3
gcsfs==2025.5.1
geomet==1.1.0
gitdb==4.0.12
GitPython==3.1.44
giturlparse==0.12.0
google-api-core==2.24.2
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.94.0
google-cloud-bigquery==3.32.0
google-cloud-bigquery-storage==2.31.0
google-cloud-core==2.4.3
google-cloud-dataproc==5.18.1
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.16.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.16.0
hexbytes==1.3.1
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanize==4.12.3
idna==3.10
importlib-metadata==6.11.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isodate==0.6.1
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
json5==0.12.0
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
keyring==25.6.0
kiwisolver==1.4.8
leather==0.4.0
locket==1.0.0
lxml==5.4.0
makefun==1.16.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mashumaro==3.14
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
more-itertools==10.7.0
msgpack==1.1.0
multidict==6.4.4
narwhals==1.41.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook_shim==0.2.4
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
ordered-set==4.1.0
orjson==3.10.18
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandas-gbq==0.29.0
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.4
partd==1.4.2
pathspec==0.12.1
pathvalidate==3.2.3
pendulum==3.1.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
ply==3.11
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
puremagic==1.29
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydata-google-auth==1.9.1
pydeck==0.9.1
Pygments==2.19.1
pylerc==4.0
pyparsing==3.2.3
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requirements-parser==0.13.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-argparse==1.7.1
rpds-py==0.24.0
rsa==4.9.1
SecretStorage==3.3.3
semver==3.0.4
Send2Trash==1.8.3
shapely==2.1.1
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowplow-tracker==1.1.0
soupsieve==2.7
sqlglot==26.21.0
sqlparse==0.5.3
stack-data==0.6.3
streamlit==1.45.1
tenacity==9.1.2
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.4.0
toml==0.10.2
tomlkit==0.13.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
truststore==0.10.1
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ujson==5.10.0
uri-template==1.3.0
urllib3==2.4.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.0
zipp==3.21.0
